requirements.txt
python-dotenv
# pip install python-dotenv

bardapi
# pip install bardapi